learning rate

+++ mhdm/tfops/callbacks.py
@@ -92,7 +92,7 @@ class NbitTreeCallback(LambdaCallback):
 			probs, code, payload, bits = self.mode(step, sample, info, tree_start, tree_end)
 
 			if tree_start:
-				points = int(info[-2])
+				points = float(info[-2])
 				bit_count = 0
 				if self.output:
 					if py7zr:
